stubgen: Fix generated stub filename when referencing out-of-directory extensions
Currently, to generate stubs for extensions not contained in the current directory,
you have to invoke stubgen like so:

`python -m nanobind.stubgen -m path.to.my.module -r -O ...`

This is for example how I do it in nanobind-bazel, because Bazel invokes scripts
from a hermetic "runfiles" directory, which is in general not the source directory.

In the recursive case, and possibly also the non-recursive case with an output
directory specified, this leads to the creation of a file called `path.to.my.pyi`,
which is not what we want (IDEs generally require that the stub file be of the
same name as the generated module).

Hence, in that case, we infer the stub file name as the last part of the module name
given to stubgen.