stubgen: Fix the detection of modules without spec.
This PR changes the logic of `is_valid_module` used in `simplify_types`.
That function uses `importlib.util.find_spec` and previously tested the
return value for `is None`. However, (1) if the spec is actually `None`,
then the function doesn't return `None` but raises `ValueError`, so that
test never succeeded and (2) if a module with a `None` spec was found,
we have still found a module, so `is_valid_module` should actually
return `True` instead of `False`.

Signed-off-by: Ingo Müller <ingomueller@google.com>