fix(cmake): Handle empty SKBUILD_SABI_VERSION
This is an empty string when it's not building for the Stable ABI.

I chose to check if it's defined and check if it's empty in two steps, though if you'd prefer, the empty check also is False if it's not defined. The form I used is slightly clearer, and might work a little better if someone passes `--warn-uninitialized` (maybe the second one passes that too, not sure).