Make stubgen compatible with sanitizers
In CMake build systems using ``nanobind_add_stub()``, a tricky issue
arises when the extension is built with sanitizers. Unless Python itself
is built with sanitizers as well, this generally requires injecting a
special sanitizer library into the Python executable at startup time.
Failing to do so causes the Python interpreter to abort when the
sanitized extension is imported. stubgen needs to run Python and import
the built extension, so this combination simply did not work so far.

This commit extends the CMake implementation of ``nanobind_add_stub()``
so that it searches the dependency list (``DEPENDS``) for any targets
that use sanitizers. It figures out the right injection library, deals
with a tricky issue on macOS (we must find out the true Python
executable, which is often hidden behind several layers of shims), and
then runs stubgen with sanitizers pre-injected.

This feature only works on macOS and Linux. I don't know enough about
sanitizers on Windows, external contributions in this regard are
welcome.