Add constructors of enum that take member definitions
The benefit of the new API is a simpler and safer implementation that
uses only the public API of the Python Enum class, mimicking how it
would be done in Python code. The exisiting API `value()` incrementally
adds member to a Enum class, entailing modifying internal states of the
Enum class.