accessor: Overload operators to allow in-place assignments on `m_base`
This enables users to write the following for in-place ops:

```cpp
nb::object obj = ...;
// increment a (mutable) counter value inline.
obj.attr("count") += nb::int(1);
```

This is useful in bindings and wrappers because it avoids the duplication of
`obj.attr()` calls. Templating the operator declaration allows their use in all
accessor implementations.

A new `test_accessor` CMake target was defined along with a Python test file,
covering the most important accessor types.