feat(stubgen): allow `__suffix__` and `__prefix__` for classes
It is currently not possible to *add* functions to a class using
patterns. This may be necessary if functions are added to a class
through mechanisms other than nanobind. This PR adds the
possibility to match on `__suffix__` and `__prefix__` within a
class, like it is possible to do within modules already, which
can be used to add anything to the beginning or end of a class using
patterns.

Signed-off-by: Ingo Müller <ingomueller@google.com>